% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginsEndsWith.r
\name{beginsEndsWith}
\alias{beginsEndsWith}
\title{Check if some text begins or ends with some other text}
\usage{
beginsEndsWith(text, find, beginSearch, caseSensitive = FALSE)
}
\arguments{
\item{text}{A character string of the text to search through.}

\item{find}{A character string of the text to search for.}

\item{beginSearch}{A logical boolean (\code{TRUE}/\code{FALSE}) that specifies if a search should be performed at the beginning of the string. \code{TRUE} = beginning, \code{FALSE} = end.}

\item{caseSensitive}{A logical boolean (\code{TRUE}/\code{FALSE}) that specifies whether text case should be matched exactly or ignored. \code{TRUE} = match case, \code{FALSE} = ignore case. Defaults to FALSE.}
}
\value{
A logical boolean (\code{TRUE}/\code{FALSE}) that states whether or not the \code{find} parameter was found within \code{text}.
}
\description{
Extends the \code{\link{substring}} function to check to see if a string begins or ends with a specified string.
}
\examples{
beginsEndsWith('test','te',TRUE) # returns TRUE
beginsEndsWith('test','test1',TRUE) # returns FALSE
beginsEndsWith('test','t',FALSE) # returns TRUE
beginsEndsWith('test','t',FALSE, TRUE) # returns FALSE

}
\seealso{
\code{\link{substring}}
}
\keyword{methods}
