% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createControlChart.R
\name{createControlChart}
\alias{createControlChart}
\title{Create Control Charts in R}
\usage{
createControlChart(data, obs, vals, stages = NA, useSD = FALSE,
  specLimit = NA, xAxisLabel = NA, yAxisLabel = NA, chartTitle = NA,
  dataFormat = "Round2")
}
\arguments{
\item{data}{A data.table of the data you want to chart. This can include extra, unneeded columns.

This will try to coerce data to a data.table if a non-data.table is passed.}

\item{obs}{A character string of the column identifying observations (often dates).}

\item{vals}{A character string of the column identifying the data points to plot.}

\item{stages}{A character string of the column identifying the stages or groups of the data points. If NA, no stages are used. Defaults to NA.}

\item{useSD}{A logical boolean (TRUE/FALSE). If TRUE, will use 3 * calculated standard deviations from the mean as the UCL/LCL. If FALSE, will use 2.66 * the average of the moving range (as is the default in Minitab) from the mean. Defaults to FALSE.}

\item{specLimit}{A one or two element vector of values specifying an upper and lower spec limit. Defaults to NA.}

\item{xAxisLabel}{A character string of the text to use as the x axis label. Defaults to NA, in which case the field name will be used.}

\item{yAxisLabel}{A character string of the text to use as the y axis label. Defaults to NA, in which case the field name will be used.}

\item{chartTitle}{A character string of the text to use as the title of the chart. Defaults to NA.}

\item{dataFormat}{A character string of the simpleFormat to be used. (See ?simpleFormat) Defaults to 'Round2'.}
}
\value{
A ggplot2 object that can be printed or added to a file/image output. 
Can also be manipulated as a normal ggplot object can by adding new layers (though this is likely unnecessary and cumbersome).
e.g. createControlChart + geom_line(aes(label = somedata))
}
\description{
Reusable code to quickly create Minitab-like control charts in R.
}
\examples{

## Creating a basic control chart

# Create a dataset for example purposes only.
my.dt<-data.table(date=seq(as.Date('2017-01-01'),as.Date('2017-04-10'),1), level=rnorm(100)/2.5)
# Alternatively: seq(as.Date('2017-01-01'),by=1,length.out=100)
my.dt[1,2] <- 3.5
my.dt[32,2] <- 2.75
my.dt[77,2] <- -3.8
my.dt[78,2] <- -3.9

# Create the control chart/run the function
createControlChart(my.dt,'date','level')
# Create same control chart with upper and lower spec limits
createControlChart(my.dt,'date','level',specLimit=c(1,-1))

## Create control chart with stages

# Create a dataset for example purposes only.
my.dt<-data.table(observation=1:100, measurement=rnorm(100)/2.5, stage=rep(1:4,each=25))
my.dt[1,2] <- 3.5
my.dt[32,2] <- 2.75
my.dt[77,2] <- -3.8
my.dt[78,2] <- -3.9

# Create the control chart/run the function
createControlChart(my.dt,'observation','measurement','stage')
# Create the same control chart with a single spec limit
createControlChart(my.dt,'observation','measurement','stage',specLimit=1)

}
\keyword{device}
\keyword{graphs}
\keyword{methods}

