% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPivotTable.R
\name{createPivotTable}
\alias{createPivotTable}
\title{Create Excel-Like Pivot Tables}
\usage{
createPivotTable(rawData, field.row, field.col, field.val, fun = "sum",
  totals.row = TRUE, totals.col = TRUE, numFrmt = "Round2", fill = "")
}
\arguments{
\item{rawData}{The formatted data to be plotted.}

\item{field.row}{A character string of the name of the field to use for the pivot row.}

\item{field.col}{A character string of the name of the field to use for the pivot column.}

\item{field.val}{A character string of the name of the field to aggregate.}

\item{fun}{A character string denoting which function to aggregate with.}

\item{totals.row}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether or not a total for rows should be included. Defaults to \code{TRUE}.}

\item{totals.col}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether or not a total for columns should be included. Defaults to \code{TRUE}.}

\item{numFrmt}{A character string of the number format to be displayed. Defaults to \code{'Round2'}.}

\item{fill}{A value to use to replace missing values (or 0 when using \code{fun='count'}) in the aggregate. Defaults to \code{NULL}.}
}
\value{
A \code{\link{data.table}} of the pivoted data.
}
\description{
Simplify the pivoting of data to print/display similiar to an Excel pivot.
}
\details{
\code{rawData} may be a \code{\link{data.table}} or \code{\link{data.frame}} (possibly a matrix as well?).
\code{fun} must be one of 'sum', 'mean', or 'count'.
The value supplied to \code{numFrmt} will be passed to \code{\link{simpleFormat}}. Any value supplied must match a valid argument for this function.
If \code{fun = 'count'}, then \code{numFrmt} is ignored. (\code{'Round0'} is used instead.)
}
\examples{
rawData <- data.table(Row = rep(LETTERS[1:5],each=10), Column = rep(LETTERS[26:22],each=5), Value = rnorm(300), Category = letters[10:14])
pivotData <- createPivotTable(rawData, field.row='Row', field.col='Column', field.val='Value')
pivotData <- createPivotTable(rawData, field.row='Row', field.col='Column', field.val='Category', fun='count', totals.row=FALSE)
pivotData <- createPivotTable(rawData, field.row='Row', field.col='Column', field.val='Value', fun='mean', numFrmt='Percent2', totals.row=FALSE, totals.col=FALSE, fill=0)
pivotData <- createPivotTable(rawData, field.row='Row', field.col='Column', field.val='Value', fun='count', fill='MISSING')

}
\seealso{
\code{\link{formatPivotTable}} \code{\link{simpleFormat}}
}
\keyword{methods}
