% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVariationChart.R
\name{createVariationChart}
\alias{createVariationChart}
\title{Create Variation Charts}
\usage{
createVariationChart(data, obs, vals, stages = NA, variationTarget = "mean",
  xAxisLabel = NA, yAxisLabel = NA, chartTitle = NA,
  dataFormat = "Round2", labelAllPoints = FALSE, dataLabelSize = 3.25,
  pos.good = TRUE)
}
\arguments{
\item{data}{A \code{\link{data.table}} of the data you want to chart.}

\item{obs}{A character string of the column identifying observations (often dates).}

\item{vals}{A character string of the column identifying the data points to plot.}

\item{stages}{A character string of the column identifying the stages or groups of the data points. If NA, no stages are used. Defaults to NA.}

\item{variationTarget}{A character string specifying the value to measure variation from. Either the 'mean', or the name of a column. Defaults to 'mean'.}

\item{xAxisLabel}{A character string of the text to use as the x axis label. Defaults to NA.}

\item{yAxisLabel}{A character string of the text to use as the y axis label. Defaults to NA.}

\item{chartTitle}{A character string of the text to use as the title of the chart. Defaults to NA.}

\item{dataFormat}{A character string format as used by \code{\link{simpleFormat}()}. Defaults to \code{'Round2'}.}

\item{labelAllPoints}{A logical boolean (\code{TRUE}/\code{FALSE}). If \code{TRUE}, will add labels to every data point on the chart. Defaults to \code{FALSE} (no labels).}

\item{dataLabelSize}{A numeric value specifying the size of data point labels. Default to 3.25. Ignored when \code{labelAllPoints} is \code{FALSE}}

\item{pos.good}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether positive values should be marked green (good/\code{TRUE}) or red (not good/\code{FALSE}). Defaults to \code{TRUE}}
}
\value{
A ggplot2 object that can be printed or added to a file/image output.
Can also be manipulated as a normal ggplot object can by adding new layers (though this is likely unnecessary and cumbersome).
e.g. createVariationChart + geom_line(aes(label = somedata))
}
\description{
Reusable code to quickly create customizable charts showing variation from a specified value.
}
\details{
\code{data} may include extra, unneeded columns. \code{createControlChart()} will try to coerce data to a \code{data.table} if not provided in this format.

Stages are only relevant when using \code{'mean'} (default) as the \code{variationTarget}. If this is not the value set, \code{stages} will be ignored. 

When \code{xAxisLabel} or \code{yAxisLabel} is NA (default), the field name from the data will be used as the axis title. If an empty string (\code{''})is passed, the axis label is completely removed.
}
\examples{

## Creating a basic variation chart, formatting the values

# Create a dataset for example purposes only.
my.dt<-data.table(date=seq(as.Date('2017-01-01'),as.Date('2017-04-10'),1), level=rnorm(100)/2.5)
# Alternatively: seq(as.Date('2017-01-01'),by=1,length.out=100)
my.dt[1,2] <- 3.5
my.dt[32,2] <- 2.75
my.dt[77,2] <- -3.8
my.dt[78,2] <- -3.9

# Create the variation chart/run the function
createVariationChart(my.dt, 'date', 'level', dataFormat = 'Round2')
# Create same variation chart with special axis/chart labels
createVariationChart(my.dt, 'date', 'level', dataFormat = 'Round2',
                   xAxisLabel = 'Date of Occurrence', yAxisLabel = 'Recording Variation',
                   chartTitle = 'Variation over Time')

## Create variation chart with stages and data labels

# Create a dataset for example purposes only.
my.dt<-data.table(observation=1:20, measurement=rnorm(20)/10, target=0.5, stage=rep(1:2,each=10))
my.dt[8,2] <- 0.9535
my.dt[17,2] <- -0.38975
my.dt[18,2] <- -0.39716

# Create the variation chart/run the function with decimal percentages
createVariationChart(my.dt, 'observation', 'measurement', 'stage',
                   labelAllPoints = TRUE, dataLabelSize = 3, dataFormat = 'Percent2')
# Create the same variation chart with whole percentages
createVariationChart(my.dt, 'observation', 'measurement', 'stage'
                   labelAllPoints = TRUE, dataLabelSize = 4, dataFormat = 'Percent0')

}
\seealso{
See \code{\link{simpleFormat}} for format options, \code{\link{TemplateHelp}} for other tools.
}
\keyword{device}
\keyword{graphs}
\keyword{methods}
