% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTextFiles.R
\name{findTextFiles}
\alias{findTextFiles}
\title{Find files containing some text.}
\usage{
findTextFiles()
}
\value{
A \code{\link{list}} of files found.
}
\description{
Wrapper/GUI for \code{\link{scanTextFiles}}, parameters for which will be gathered through a user form.
}
\details{
When saving a text file of results, the file (\code{FoundFiles.txt}) will be saved in the path on the selection form.
If not saving results, the user form is closed/cancelled, or no results are found, then \code{NA} is returned by the function instead of a list.
}
\examples{
findTextFiles() # Return a list of files
findTextFiles(TRUE) # Save a text file with the list instead.
}
\seealso{
\code{\link{scanTextFiles}}
}
\keyword{methods}
