% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPivotTable.R
\name{formatPivotTable}
\alias{formatPivotTable}
\title{Format Excel-Like Pivot Tables}
\usage{
formatPivotTable(pvData, pvTitle = NA, pvSubT = NA, txtScale = 0.75,
  hdrScale = 1, ttlScale = 1, subTtlScale = 1, width = NA,
  height = NA)
}
\arguments{
\item{pvData}{The formatted data to be plotted.}

\item{pvTitle}{A optional character string to use as the table's overall title.}

\item{pvSubT}{A optional character string to use as the table's subtitle.}

\item{txtScale}{A numeric value representing the relative size of the values in the table.}

\item{hdrScale}{A numeric value representing the relative size of the column headers.}

\item{ttlScale}{A numeric value representing the relative size of title (if supplied).}

\item{subTtlScale}{A numeric value representing the relative size of the subtitle (if supplied).}

\item{width}{A numeric value representing the width (in inches) the final table should fill.}

\item{height}{A numeric value representing the height (in inches) the final table should fill.}
}
\value{
A \code{\link{tableGrob}} prepared to print to pdf/image/other output format.
}
\description{
Format pivoted data to print/display similiar to an Excel pivot.
}
\details{
\code{pvData} may be a \code{\link{data.table}} or \code{\link{data.frame}} (possibly a matrix as well?).
It is not necessary include \code{pvTitle} when using \code{pvSubT}, as the function will format according to whichever arrangement is presented. However, scaling for both is dependent upon the explict argument given (\code{ttlScale} and \code{subTtlScale}, respectively.). However, both used together is also acceptable.
If \code{width} and/or \code{height} are omitted, the table will be sized to fit the data.
}
\examples{
pTable <- formatPivotTable(pvData,'TITLE','Subtitle')
pTable <- formatPivotTable(pvData,'TITLE', ttlScale = 2)
pTable <- formatPivotTable(pvData, txtScale = 0.5, hdrScale = 0.75)

}
\seealso{
\code{\link{createPivotTable}}
}
\keyword{methods}
