% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevWeekday.r
\name{getPrevWeekday}
\alias{getPrevWeekday}
\title{Get Previous Weekday}
\usage{
getPrevWeekday(date)
}
\arguments{
\item{date}{Origin date from which to find the previous date.}
}
\value{
A date value of the preceeding weekday.
}
\description{
Returns the previous weekday based on a date provided to the function.
}
\details{
Date provided must be formatted as a valid R Date. This function will not attempt to convert.
The date value returned will typically be \code{date - 1}, however, in teh case of running on a Saturday, Sunday, or Monday, the previous Friday date is returned.
}
\examples{
beginsEndsWith('test','te',TRUE) # returns TRUE
beginsEndsWith('test','test1',TRUE) # returns FALSE
beginsEndsWith('test','t',FALSE) # returns TRUE
beginsEndsWith('test','t',FALSE, TRUE) # returns FALSE

}
\seealso{
\code{\link{as.Date}}
}
\keyword{date}
\keyword{methods}
