% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserDates.r
\name{getUserDates}
\alias{getUserDates}
\title{Date Selection}
\usage{
getUserDates(numFields = 2, outputFormat = "R")
}
\arguments{
\item{numFields}{A \code{\link{numeric}} value specifying the number of date fields to prompt. Defaults to \code{2}.}

\item{outputFormat}{A \code{\link{character}} string specifying format type of the output. Defaults to \code{'R'}. See details for more info.}
}
\value{
A vector containing the formatted dates.
}
\description{
Provide a user form to prompt the user for date(s), which will then return the inputted values in various formats.
}
\details{
\code{numFields} must be between 1 and 4, inclusive. Each corresponding field on the form will be labeled "Date x", where x is each integer within the range provided.
The default value is 2, (\code{Date 1}, \code{Date 2}) which generally allows for the selection of a start and end date, but this need not necessarily be used in this way.

\code{outputFormat} must be one of:
\itemize{
\item{\code{'R'} - Default. Will return \code{Date} objects. e.g. \code{as.Date('2018-07-15')}}
\item{\code{'Print'} - Will return strings, formatted for normal reading. e.g. \code{07/15/2018}}
\item{\code{'SQL'} - will return strings, formatted to be used in SQL queries. e.g. \code{'15-JUL-2018'} (note the quotes)}
}
}
\examples{

dates <- getUserDates()
dates <- getUserDates(1, 'Print')
dates <- getUserDates(numFields = 3, 'SQL')
}
\seealso{
\code{\link{fgui::gui}}, \code{\link{format.Date}}
}
\keyword{interface}
\keyword{io}
\keyword{methods}
