% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserPass.r
\name{getUserPass}
\alias{getUserPass}
\title{Collect Username and Masked Password}
\usage{
getUserPass(helperLabel = NA, helperUsername = NA, helperPassword = NA)
}
\arguments{
\item{helperLabel}{A character string stating the name of the resource to connect to.}

\item{helperUsername}{A character string of the label to prompt for the username.}

\item{helperPassword}{A character string of the label to prompt for the password.}
}
\value{
A vector containing two character strings. [1] = username; [2] = password
}
\description{
Reusable code to more securely accept usernames/passwords, rather than having to hard-code this information.
Typically used for logging in to the database, though this could apply to any other uses as well.
}
\examples{
# Base usage
loginInfo <- getUserPass()
username <- loginInfo[1]
userpass <- loginInfo[2]
# Custom labels
loginInfo <- getUserPass(helperLabel = 'My Database', 
                         helperUsername = 'Enter login for My Database:',
                         helperPassword = 'Enter password for My Database:')
username <- loginInfo[1]
userpass <- loginInfo[2]
}
\keyword{IO}
\keyword{methods}
