% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateLatestIterationICNs.r
\name{isolateLatestIterationICNs}
\alias{isolateLatestIterationICNs}
\title{Isolate Latest Iteration ICNs}
\usage{
isolateLatestIterationICNs(data, full = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{A \code{\link{data.table}} containing the data to filter.}

\item{full}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether the returned result should include all data (\code{TRUE}) or only the ICNs (\code{FALSE}). Defaults to \code{TRUE}.}

\item{verbose}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether output messages should be displayed. Defaults to \code{TRUE}.}
}
\value{
A \code{\link{data.table}} containing the filtered data. Full dataset if \code{full = TRUE}, single column of ICNs if \code{full = FALSE}.
}
\description{
Given a listing of bill data (must include ICN), group all related ICN bases and return all data for only those that are the highest iteration.
}
\details{
\code{isolateLatestIterationICNs()} will try to coerce data to a \code{data.table} if not provided in this format.
}
\examples{
library (RTemplates)
sql <- "SELECT ICN, FNLZD_PAID_POSTING_DT 
        FROM BILLREVW.BILL_HDR
        WHERE FNLZD_PAID_POSTING_DT BETWEEN '01-JAN-2016'
           AND '31-DEC-2016'
        AND BILLREVW_BE_ID = '008'"

dt <- runSQLQuery(sql,"wcdw64",verbose=TRUE)
isolateLatestIterationICNs(dt)

}
\keyword{methods}
