% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTextFromFile.r
\name{loadTextFromFile}
\alias{loadTextFromFile}
\title{Read a text file}
\usage{
loadTextFromFile(filename, stripEscape = FALSE)
}
\arguments{
\item{filename}{A character string of the name of the file to read. If the full path is not provided, R will use the current working directory.}

\item{stripEscape}{A boolean indicating whether or not to remove special/escape characters (\code{\\n})form the read text. Defaults to \code{FALSE}.}
}
\value{
A character string of the contents of the file read.
}
\description{
Read in a text file (.txt, .sql, etc.) and parse the text to a character string.
}
\examples{
myQuery <- loadTextFromFile('query.sql')

}
\seealso{
\code{\link{readChar}}
}
\keyword{io}
\keyword{methods}
