% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locationTiming.r
\name{locationTiming}
\alias{locationTiming}
\title{Generate Location Timings}
\usage{
locationTiming(locData, ICN = "ICN", Location = "SYST_LCTN_ID",
  TimeStamp = "LAST_ACTN_TMSTAMP", units = "days")
}
\arguments{
\item{locData}{A \code{\link{data.table}} containing the data to parse.}

\item{ICN}{A \code{\link{character}} vector/string specifying the name of the ICN field. Defaults to DB-standard 'ICN'.}

\item{Location}{A \code{\link{character}} vector/string specifying the name of the location field. Defaults to DB-standard 'SYST_LCTN_ID'.}

\item{TimeStamp}{A \code{\link{character}} vector/string specifying the name of the timestamp field. Defaults to DB-standard 'LAST_ACTN_TMSTAMP'.}

\item{units}{A \code{\link{character}} vector/string indicating the units/format for the output resulting TAT. Defaults to 'days'.}
}
\value{
A \code{\link{data.table}} or \code{\link{list}} containing the output of the queries.
}
\description{
Parse and group item history to get total time per individual instance of group activity (rather than grouping across all instances of a group as a single unit).
}
\details{
\code{locData} must contain at least three columns (names can vary):
\itemize{
\item{\code{ICN}}
\item{\code{Location}}
\item{\code{TimeStamp}}
}
Other columns may be included, but will be stripped from the data and not returned.

See \code{\link{seq.Date}} for further information about the \code{units} parameter.
}
\examples{

dt <- runSQLQuery('History.sql', DB='DBName',
            verbose = TRUE)
dtLocTime <- locationTiming(dt)
dtLocTimeAlt <- locationTiming(
  dt,
  ICN = 'Control_Number',
  Location = 'Location', 
  TimeStamp = 'Timestamp',
  units = 'hours')
}
\seealso{
\code{\link{difftime}} \code{\link{merge}}
}
\keyword{methods}
