% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.console.r
\name{progress.console}
\alias{progress.console}
\title{Display a progress bar in the console window}
\usage{
progress.console(progCurrent, progTotal, dispLength = 30,
  showPercent = TRUE, appendMsg = NA)
}
\arguments{
\item{progTotal}{A numeric value giving the total step}

\item{dispLength}{An integervalue that specifies how many characters to expand the prgress bar to. Defaults to 30.}

\item{showPercent}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether or not a percentage should be displayed along side the raw progress count. Defaults to \code{TRUE}.}

\item{appendMsg}{A character vector of additional informational text to append to the end of the progress bar. Optional.}

\item{progCurent}{A numeric value giving the current step of the progress.}
}
\value{
Prints the progress bar using text to the console window.
}
\description{
Shows a progress bar on the text console that updates as needed.
}
\examples{
# Basic Example:
totalSteps <- 25
for(i in 1:totalSteps) {
   progress.console(i,totalSteps)
   Sys.sleep(.5)
}

# More Features:
totalSteps <- 37
for(i in 1:totalSteps) {
   progress.console(progCurrent = i, progTotal = totalSteps,
      dispLength = 45, showPercent = FALSE,
      appendMsg = paste0(' - Other ','messages as needed - ',i)
   )
   Sys.sleep(.5)
}
}
\keyword{methods}
\keyword{text}
