% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemplates-package.R
\docType{package}
\name{rtemplates-package}
\alias{rtemplates}
\alias{rtemplates-package}
\title{rtemplates: Personal templates for rmarkdown.}
\description{
Personal templates for rmarkdown and sweave.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jolars/rtemplates}
  \item Report bugs at \url{https://github.com/jolars/rtemplates/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{mail@larssonjohan.com} (0000-0002-4029-5945)

}
\keyword{internal}
