% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDataToTempTable.r
\name{saveDataToTempTable}
\alias{saveDataToTempTable}
\title{Store data in a temporary database table}
\usage{
saveDataToTempTable(DBTo, data, tempName = "TEMP", index = NULL,
  verbose = FALSE)
}
\arguments{
\item{DBTo}{A character string of the name of the ODBC database connection to send data to.}

\item{data}{An R dataset (\code{\link{data.frame}} or \code{\link{data.table}}) to stop in the temporary table.}

\item{tempName}{A character string of the name to use for the temporary table. Defaults to \code{'TEMP'}.}

\item{index}{A named list containing information about how to create indicies. See details for more information.}

\item{verbose}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether output messages should be displayed. Defaults to \code{FALSE}.}
}
\value{
Result will be a table in the requested database with a name given as defined in the details section.
}
\description{
Wrapper to store data in a temporary table. Data can be an R dataset or pulled directly from the destination or external source database.
}
\details{
You must have appropriate read/write access to the database identified in \code{DBTo} in order for this funciton to work.

The table created in the database will be named \code{<username>.<tempName>} if \code{tempName} supplied, otherwise, \code{<username>.temp}

The named list supplied to the \code{index} argument should be a vector of column names, with each named element being the intended name of the index. If left blank, no indexes will be produced. The named list should be constructed as follows:
\itemize{
  \item TYPE: Named lists with names \code{'UNIQUE'} and/or \code{'NORMAL'} noting the type of index to create.
  \item NAME: Named lists with names to be used as the index name. The index will be named according to the table-naming scheme above: \code{<table_name>_<index_name>}.
  \item COLUMNS: Character/String elemements corresponding to the column names to be used in the index.
}
See the examples section for a correct, working example of an index list.
}
\examples{

myODBCDatabase <- 'ODBCConName'
myData <- data.table(A=1:100000, B=LETTERS[1:10], C=rnorm(100000))
saveDataToTempTable(DBTo = myODBCDatabase, data = myData)
# With indexing:
myIndex <- list( # Top level name indicates the type (UNIQUE/NORMAL)
                UNIQUE = list( # Second level name indicates the name of the index
                              KEY = 'A'),
                 # Top level name indicates the type (UNIQUE/NORMAL)
                NORMAL = list( # Second level name indicates the name of the index
                             ALT = c('B','C'))
               )
saveDataToTempTable(myODBCDatabase, myData, tempName = 'MY_TEMP_TABLE', index = myIndex, verbose = TRUE)
}
\keyword{connections}
\keyword{database}
\keyword{interface}
\keyword{io}
\keyword{methods}
