% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanTextFiles.r
\name{scanTextFiles}
\alias{scanTextFiles}
\title{Find files containing some text.}
\usage{
scanTextFiles(srchText, path = NA, ignoreCase = TRUE, subdirs = FALSE,
  extension = NA)
}
\arguments{
\item{srchText}{A character string of the text to search for.}

\item{path}{A valid path that contains files to search through.}

\item{ignoreCase}{A logical boolean (\code{TRUE}/\code{FALSE}) that specifies whether text case should be matched exactly or ignored. Defaults to \code{TRUE}.}

\item{subdirs}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether or not subdirectories of \code{path} should be scanned. Defaults to\code{FALSE}.}

\item{extension}{A character vector of extensions to limit file search.}
}
\value{
A list of files contatining the value assigned to \code{srchText}.
}
\description{
Reads through files in a directory and returns a list of the text files found that contain a specified text string.
}
\details{
If \code{path} is not specified, R's working directory (\code{\link{getwd}}) will be used.

When specifying file types/extensions with the \code{extension} argument, you may pass a single string (\code{'.txt'}) or a list of multiple options (\code{c('.txt','.csv')})
Extension strings need not have the preceding period; either with or without will work. i.e. \code{'.txt'} is equivalent to \code{'txt'}.
}
\examples{
scanTextFiles('default parameters')
scanTextFiles('specific path', path = 'P:/test/') # Note the direction of slashes.
scanTextFiles('case sensitive, all subdirectories, specific extensions', ignoreCase = FALSE, subdirs = TRUE, extention = c('.txt','csv','sql','.r'))
}
\seealso{
\code{\link{loadTextFromFile}}, \code{\link{list.files}}
}
\keyword{methods}
\keyword{text}
