% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleFormat.r
\name{simpleFormat}
\alias{simpleFormat}
\title{Simplified formatting option for common formats}
\usage{
simpleFormat(value, formatType)
}
\arguments{
\item{value}{A numeric value to be formatted.}

\item{formatType}{A character string of the type to be used.}
}
\value{
A characer string of the value formatted as specified.
}
\description{
Formats numbers in a variety of common formats without having to learn/understand the finer details of formatting in R.
}
\details{
Valid options for \code{formatType} are:
\itemize{
  \item 'Round0' = \code{format(round(value,0),nsmall=0,scientific=FALSE)} # Rounding to whole numbers
  \item 'Round2' = \code{format(round(value,2),nsmall=2,scientific=FALSE)} # Rounding to two decimal places
  \item 'Percent0' = \code{paste0(formatC(round(value*100,0),format='f',digits=0),'\%')} # Whole percent
  \item 'Percent2' = \code{paste0(formatC(round(value*100,2),format='f',digits=2),'\%')} # Percent to two decimal places
  \item 'Comma0' = \code{format(round(value,0),big.mark=',',nsmall=0,scientific=FALSE)} # Rounding to whole numbers with thousands separator
  \item 'Comma2' = \code{format(round(value,2),big.mark=',',nsmall=2,scientific=FALSE)} # Rounding to two decimal places with thousands separator
  \item 'Currency0' = \code{paste0('$',format(round(value,0),big.mark=',',nsmall=0,scientific=FALSE))} # Currency with no cents
  \item 'Currency2' = \code{paste0('$',format(round(value,2),big.mark=',',nsmall=2,scientific=FALSE))} # Currency with cents
  \item 'CurrencyK' = \code{paste0('$',format(round(value/1000,0),big.mark=',',nsmall=0,scientific=FALSE),' K')} # Currency in thousands of $
  \item 'CurrencyM' = \code{paste0('$',format(round(value/1000000,0),big.mark=',',nsmall=0,scientific=FALSE),' mil')} # Currency in millions of $
}
}
\examples{

simpleFormat(12.3456789, 'Round0') # returns '12'
simpleFormat(12.3456789, 'Round2') # returns '12.35'
simpleFormat(123456789, 'Comma2') # returns '123,456,789.00'
simpleFormat(0.9987654, 'Percent2') # returns '99.88\%'
simpleFormat(12345678.90, 'Currency2') # returns '$12,345,678.90'
simpleFormat(12345678.90, 'CurrencyK') # returns '$12,346 K'
simpleFormat(12345678.90, 'CurrencyM') # returns '$12 mil'

}
\keyword{methods}
