% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAT.r
\name{TAT}
\alias{TAT}
\title{Calculate Time Between Two Dates}
\usage{
TAT(startDate = NA, endDate = NA, excludeWeekends = TRUE,
  excludeHolidays = TRUE, specialExclusions = NA, sameDayCounts = TRUE,
  calcDays = TRUE, dayStartTime = NA, dayEndTime = NA)
}
\arguments{
\item{startDate}{A \code{\link{Date}} or datetime (\code{\link{POSIXct}}) (or a vector of same) specifying the earliest date/time in the period.}

\item{endDate}{A \code{\link{Date}} or datetime (\code{\link{POSIXct}}) (or a vector of same) specifying the latest date/time in the period.}

\item{excludeWeekends}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether weekend dates should be counted (\code{FALSE}) or excluded (\code{TRUE}). Defaults to \code{TRUE}.}

\item{excludeHolidays}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether holiday dates should be counted (\code{FALSE}) or excluded (\code{TRUE}). Defaults to \code{TRUE}.}

\item{specialExclusions}{A vector of \code{\link{Date}} values to be excluded. Can overlap weekend/default holidays.}

\item{sameDayCounts}{A logical boolean (\code{TRUE}/\code{FALSE}) stating whether same-day spans should be counted as 0 (\code{FALSE}) or 1 (\code{TRUE}). Defaults to \code{TRUE}.}

\item{calcDays}{<TO BE IMPLEMENTED> A logical boolean (\code{TRUE}/\code{FALSE}) stating whether the value returned should be in days (\code{TRUE}) or time (\code{FALSE}). Defaults to \code{TRUE}.}

\item{dayStartTime}{<TO BE IMPLEMENTED>}

\item{dayEndTime}{<TO BE IMPLEMENTED>}
}
\value{
A \code{\link{numeric}} value, giving the number of days, or a time value.
}
\description{
Calculates the amount of time between two date/timestamps, reported in days or time with options for special exclusions.
}
\details{
Calculates turn around time between two dates, either as whole-day numbers or (not yet implemented) partial days/hours.

If dates are not provided to the function, a pop up will appear asking for the values to be entered.
}
\examples{
TAT()
#Varies, depending upon user input.
TAT(as.Date('2017-07-01'),as.Date('2017-07-31'))
#19
TAT(as.Date('2017-07-01'),as.Date('2017-07-31'), excludeWeekends = FALSE)
#29
TAT(as.Date('2017-07-01'),as.Date('2017-07-31'), excludeWeekends = FALSE, excludeHolidays = FALSE)
#31
TAT(as.Date('2017-07-01'),as.Date('2017-07-31'), excludeHolidays = FALSE)
#21
#Run for multiple ranges. Good for getting TAT for a series of items.
sDates <- c(as.Date('2018-01-01'), as.Date('2018-02-01'))
eDates <- c(as.Date('2018-02-01'), as.Date('2018-03-01'))
TAT(sDates, eDates)
#23, 21
}
\keyword{methods}
