% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLog.r
\name{writeLog}
\alias{writeLog}
\title{Write Log to Console}
\usage{
writeLog(text, inc.date = TRUE)
}
\arguments{
\item{text}{A character string of the text to post.}
}
\value{
Writes \code{text} to console with the current timestamp.
}
\description{
Extends the \code{\link{writeLines}} function to add a timestamp when logging.
}
\examples{
writeLog('Testing log.')

}
\seealso{
\code{\link{writeLines}}
}
\keyword{methods}
